♻️ refactor: Separating a function from the product.types.ts file to product.utils.ts file - Modify the paths of other files that got error during the change as well
