+++ src/app/(purchase)/cart/_components/CartClient.tsx
@@ -7,8 +7,8 @@ import { Checkbox } from '@/components/ui/Checkbox';
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/Dialog';
 import { removeFromCartAction, updateCartOptionAction, updateCartQuantityAction } from '@/lib/actions/cartServerActions';
 import { createCartPurchaseTempOrderAction } from '@/lib/actions/order/orderServerActions';
+import { getImageUrl } from '@/lib/utils/product.utils';
 import { CartItem } from '@/types/cart.types';
-import { getImageUrl } from '@/types/product.types';
 import { Trash2 } from 'lucide-react';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
