+++ src/app/my-page/bookmarks/products/_components/AddToCartDialog.tsx
@@ -10,7 +10,8 @@ import { Button } from '@/components/ui/Button';
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/Dialog';
 import { addToCartAction } from '@/lib/actions/cartServerActions';
 import { getProductById } from '@/lib/functions/shop/productClientFunctions';
-import { getProductPotColors, Product } from '@/types/product.types';
+import { getProductPotColors } from '@/lib/utils/product.utils';
+import { Product } from '@/types/product.types';
 import { Minus, Plus } from 'lucide-react';
 import Image from 'next/image';
 import { useCallback, useEffect, useMemo, useState } from 'react';
